python_server/pyproject.toml
[project]
name = "et"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "langchain>=1.2.13",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
    "websockets>=15.0",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.0",
    "pillow>=11.2.0",
    "langchain-google-genai>=2.1.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "langchain-text-splitters>=1.1.1",
]
